발전량 조회, 발전 수익 계산 페이지 기능 완성 및 server 구축 완료

+++ BackEnd/src/main/java/com/example/slsl_server/model/Power.java
@@ -16,5 +16,5 @@ public class Power {
     @Column(columnDefinition="TEXT")
     private String loc;
     private LocalDate tm;
-    private Long value;
+    private double value;
 }
